--- app/containers/BookSlot.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const BookSlot = () => {
-	return (
-		<div>Book a slot </div>
-	)
-}
-
-export default BookSlot;
